+++ Dropshiping.BackEnd.Project/Dropshiping.BackEnd.Services/UserServices/Implementation/UserOrderService.cs
@@ -1,4 +1,5 @@
 ﻿using Dropshiping.BackEnd.DataAccess.Interface;
+using Dropshiping.BackEnd.Domain.ProductModels;
 using Dropshiping.BackEnd.Domain.UserModels;
 using Dropshiping.BackEnd.Dtos.UserOrderDtos;
 using Dropshiping.BackEnd.Mappers.UserOrderMappers;
@@ -9,9 +10,13 @@ namespace Dropshiping.BackEnd.Services.UserServices.Implementation
     public class UserOrderService : IUserOrderService
     {
         private IRepository<UserOrder> _userOrderRepository;
-        public UserOrderService(IRepository<UserOrder> userOrderRepository)
+        private IUserRepository _userRepository;
+        private IRepository<Order> _orderRepository;
+        public UserOrderService(IRepository<UserOrder> userOrderRepository, IUserRepository userRepository, IRepository<Order> orderRepository)
         {
             _userOrderRepository = userOrderRepository;
+            _userRepository = userRepository;
+            _orderRepository = orderRepository;
         }
 
         public List<UserOrderDto> GetAll()
@@ -34,6 +39,19 @@ namespace Dropshiping.BackEnd.Services.UserServices.Implementation
 
         public void Add(string userId, string orderId)
         {
+            var user = _userRepository.GetById(userId);
+            var order = _orderRepository.GetById(orderId);
+
+            if (user == null)
+            {
+                throw new KeyNotFoundException($"User with id {userId} doesn't exists.");
+            }
+
+            if (order == null)
+            {
+                throw new KeyNotFoundException($"Order with id {orderId} doesn't exists.");
+            }
+
             var userOrder = UserOrderMapper.ToUserOrderDomain(userId, orderId);
             _userOrderRepository.Add(userOrder);
         }
@@ -51,16 +69,12 @@ namespace Dropshiping.BackEnd.Services.UserServices.Implementation
 
         public void DeleteById(string id)
         {
-            //var userOrder = GetById(id);
+            var userOrder = GetById(id);
 
-            //if (userOrder.Id == null)
-            //{
-            //    throw new KeyNotFoundException($"UserOrder with id {id} was not found.");
-            //}
-            //if (id == "")
-            //{
-            //    throw new ArgumentException("You must enter id");
-            //}
+            if (userOrder == null)
+            {
+                throw new KeyNotFoundException($"UserOrder with id {id} was not found.");
+            }
 
             _userOrderRepository.Delete(id);
         }
